airflow/requirements.txt
astro-sdk-python==1.6.2